Add missing methods to star the application

+++ blockchain/src/main/java/com/hackathon/blockchain/model/Asset.java
@@ -22,7 +22,7 @@ public class Asset {
     private String symbol;
     private Double quantity;
 
-    private Double purchasedPrice;
+    private Integer purchasedPrice;
 
     @ManyToOne
     private Wallet wallet;
